Create override mechanism in ACL
For the github-oauth-plugin I need the ability to support existing authorization strategies with some extras.

I don't want to fork the existing functionality but rather decorate their chosen AuthorizationStrategy with some pre-checks.

This commit adds an override to the ACL class that will delegate to the ACLPermissionOverride extension first before applying the implemented ACL permission checks.